test: create tests for user service class

+++ app/Services/UserService.php
@@ -3,15 +3,13 @@
 namespace App\Services;
 
 use App\Http\Repositories\UserRepository;
-use App\Http\Resources\Task\TaskResource;
 use App\Http\Resources\User\UserResource;
 use App\Jobs\MailJob;
-use App\Mail\Notification;
-use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+use Illuminate\Contracts\Queue\EntityNotFoundException;
 use Illuminate\Support\Facades\Log;
-use Illuminate\Support\Facades\Mail;
 use Tymon\JWTAuth\Exceptions\JWTException;
 use Tymon\JWTAuth\Facades\JWTAuth;
+use Illuminate\Support\Facades\Hash;
 
 class UserService {
 
@@ -26,7 +24,9 @@ class UserService {
     }
 
     public function store(mixed $data): UserResource {
-        $data['password'] = bcrypt($data['password']);
+        $data['password'] = Hash::make($data['password']);
+
+        Log::info($data);
 
         $user = $this->repository->store($data);
 
@@ -49,8 +49,12 @@ class UserService {
     }
 
     public function update(int $id, mixed $data): UserResource {
+        if($this->repository->existsByColumn('id', $id) === false) {
+            throw new EntityNotFoundException('User', $id);
+        }
+
         $user = $this->repository->show($id);
-        $data['password'] = bcrypt($data['password']);
+        $data['password'] = Hash::make($data['password']);
 
         return new UserResource($this->repository->update($user, $data));
     }
@@ -68,6 +72,10 @@ class UserService {
     }
 
     public function destroy(int $id): void {
+        if($this->repository->existsByColumn('id', $id) === false) {
+            throw new EntityNotFoundException('User', $id);
+        }
+
         $this->repository->destroy($id);
     }
 }
